Fixes to meet maintainers specs
