+++ src/providers/index.js
@@ -1,0 +1,11 @@
+import githubConfig from './github';
+import bitbucketConfig from './bitbucket';
+import azureConfig from './azure';
+
+const providerConfig = {
+  github: githubConfig,
+  bitbucket: bitbucketConfig,
+  azure: azureConfig,
+};
+
+export default providerConfig;
